+++ src/components/controllable.rs
@@ -1,12 +1,12 @@
 use super::ComponentTrait;
 
 pub struct Controllable {
-  owner: u32,
+  owner: usize,
   is_active: bool,
 }
 
 impl ComponentTrait for Controllable {
-  fn new(entity_id: u32) -> Self {
+  fn new(entity_id: usize) -> Self {
     Self {
       owner: entity_id,
       is_active: true,
