+++ src/components/visible.rs
@@ -1,13 +1,13 @@
 use super::ComponentTrait;
 
 pub struct Visible {
-  pub owner: u32,
+  pub owner: usize,
   pub is_visible: bool,
   pub sprite: char,
 }
 
 impl ComponentTrait for Visible {
-  fn new(entity_id: u32) -> Self {
+  fn new(entity_id: usize) -> Self {
     Self {
       owner: entity_id,
       is_visible: false,
