+++ src/engine/mod.rs
@@ -6,7 +6,7 @@ pub struct GameEngine {
   pub entities: EntityManager,
   pub components: ComponentManager,
 
-  pub active_camera_id: Option<u32>,
+  pub active_camera_id: Option<usize>,
 }
 
 impl GameEngine {
@@ -19,7 +19,7 @@ impl GameEngine {
     }
   }
 
-  pub fn set_active_camera(&mut self, camera_id: u32) -> Result<(), String> {
+  pub fn set_active_camera(&mut self, camera_id: usize) -> Result<(), String> {
     let camera_error = Screen::is_valid_camera(self, camera_id);
     if let Err(error_message) = camera_error {
       return Err(error_message);
