+++ src/factories/brush.rs
@@ -3,7 +3,7 @@ use crate::GameEngine;
 
 pub struct BrushFactory {}
 impl Factory for BrushFactory {
-  fn new(engine: &mut GameEngine) -> Result<u32, String> {
+  fn new(engine: &mut GameEngine) -> Result<usize, String> {
     let brush_id = engine.entities.add("brush".to_string());
 
     engine.components.translations.register(brush_id);
