+++ src/factories/camera.rs
@@ -8,7 +8,7 @@ impl Factory for CameraFactory {
   /// 
   /// Creates a new camera for the given Game Engine and returns the ID. If the
   /// engine does not have an active camera, this camera is set as the active camera
-  fn new(engine: &mut GameEngine) -> Result<u32, String> {
+  fn new(engine: &mut GameEngine) -> Result<usize, String> {
     let camera_id = engine.entities.add("camera".to_string());
 
     engine.components.translations.register(camera_id);
